v0.3.4 prep 1
- Started integrating methods for guild and channel data fetch
- Added new methods (DELETE, PUT) into req handler
- Changed req handler to sued promise based system (guarenteed reliabikity when data is sent back to client)
- Started prepping for new embed methods
