+++ src/client/rest/ReqHandler.ts
@@ -9,29 +9,70 @@ export class ReqHandler {
 
   constructor(private client: Client) {
   }
+
+  // GET DATA FROM ROUTE
   async get(route: string): Promise<any> {
-    const request: Request = new Request(route, {
-      method: "GET",
-      client: this.client,
+    return new Promise((resolve, reject) => {
+      const request: Request = new Request(route, {
+        method: "GET",
+        client: this.client,
+        resolve,
+        reject,
+      });
+      if (!this.queue.has(route)) {
+        this.queue.set(route, new ReqQueue(this.client));
+      }
+      let res = this.queue.get(route)?.push(request);
     });
-    if (!this.queue.has(route)) {
-      this.queue.set(route, new ReqQueue(this.client));
-    }
-    let res = await this.queue.get(route)?.push(request);
-    return res;
   }
 
+  // POST TO ROUTE
   async post(route: string, body?: object | string) {
-    const request: Request = new Request(route, {
-      method: "POST",
-      body,
-      client: this.client,
+    return new Promise((resolve, reject) => {
+      const request: Request = new Request(route, {
+        method: "POST",
+        body,
+        client: this.client,
+        resolve,
+        reject,
+      });
+      if (!this.queue.has(route)) {
+        this.queue.set(route, new ReqQueue(this.client));
+      }
+      this.queue.get(route)?.push(request);
     });
-    if (!this.queue.has(route)) {
-      this.queue.set(route, new ReqQueue(this.client));
-    }
-    this.queue.get(route)?.push(request);
   }
 
-  // Add DELETE and PATCH methods for changing channel/guild data, and removing channels etc
+  // PATCH UP ROUTE
+  async patch(route: string, body?: object | string) {
+    return new Promise((resolve, reject) => {
+      const request: Request = new Request(route, {
+        method: "PATCH",
+        body,
+        client: this.client,
+        resolve,
+        reject,
+      });
+      if (!this.queue.has(route)) {
+        this.queue.set(route, new ReqQueue(this.client));
+      }
+      this.queue.get(route)?.push(request);
+    });
+  }
+
+  // DELETE DATA FROM ROUTE
+  async delete(route: string) {
+    return new Promise((resolve, reject) => {
+      const request: Request = new Request(route, {
+        method: "DELETE",
+        client: this.client,
+        resolve,
+        reject,
+      });
+      if (!this.queue.has(route)) {
+        this.queue.set(route, new ReqQueue(this.client));
+      }
+      this.queue.get(route)?.push(request);
+    });
+  }
 }
